Working leaderboard
Users are now redirected to a leaderboard after game over. Displayed as
a table of top 10 scores and the score of the game that was just
played. Made changes to CSS files to accommodate the ew table added but
the table still requires further styling. Added a .gitignore file and
made changes to workspace.xml